Выполнил разбиение на модули
